Déplacement des unités et correction d'un problème de génération de carte (non aléatoire)

+++ SmallWorld/Units/Unit.cs
@@ -82,7 +82,7 @@ namespace PetitMonde.Units
         /// <summary>
         /// Current moving points of this unit
         /// </summary>
-        float MovingPoints
+        double MovingPoints
         {
             get;
             set;
